feat(core): leverage go-cron v0.6.0 new features
Upgrade go-cron to v0.6.0 and implement new scheduler capabilities:

- Graceful Shutdown: Add StopWithTimeout() and StopAndWait() methods
  for controlled scheduler termination with job completion waiting
- Job Naming: Use WithName() for O(1) job lookup via cron.EntryByName()
- Job Tagging: Add AddJobWithTags(), RemoveJobsByTag(), GetJobsByTag()
  for logical job grouping and batch operations
- ObservabilityHooks: Integrate JobStarted, JobCompleted, JobScheduled
  callbacks with MetricsRecorder for comprehensive job monitoring
- Fix DisableJob/EnableJob: Handle go-cron's async removal with
  exponential backoff polling in EnableJob
- Update stress tests: Add staggered worker starts and reduced
  contention to prevent timeouts

This completes the migration to go-cron v0.6.0, taking advantage of
all major new features while maintaining backward compatibility.

+++ core/bare_job.go
@@ -25,7 +25,7 @@ type BareJob struct {
 	lock    sync.Mutex
 	history []*Execution
 	lastRun *Execution
-	cronID  int
+	cronID  uint64
 }
 
 func (j *BareJob) GetName() string {
@@ -52,11 +52,11 @@ func (j *BareJob) NotifyStop() {
 	atomic.AddInt32(&j.running, -1)
 }
 
-func (j *BareJob) GetCronJobID() int {
+func (j *BareJob) GetCronJobID() uint64 {
 	return j.cronID
 }
 
-func (j *BareJob) SetCronJobID(id int) {
+func (j *BareJob) SetCronJobID(id uint64) {
 	j.cronID = id
 }
